chore: apply suggestions from CR
